feat: Se agrego lógica de reservaciones para residentes

+++ Colonia/src/main/java/Control/Colonia/Controller/AmenidadController.java
@@ -19,37 +19,48 @@ public class AmenidadController {
     @Autowired
     private AmenidadService amenidadService;
 
-    @GetMapping
-    public String mostrarAmenidades(HttpSession session, Model model) {
-        List<Amenidad> lista = amenidadService.getAllAmenidad();
-        model.addAttribute("amenidades", lista);
-        model.addAttribute("amenidadForm", new Amenidad());
-
+    private boolean cargarDatos(HttpSession session, Model model){
         String tipo = (String) session.getAttribute("tipoUsuario");
 
-        if (tipo == null) {
+        if (tipo == null){
+            return false;
+        }
+
+        model.addAttribute("amenidades", amenidadService.getAllAmenidad());
+
+        return true;
+    }
+
+    @GetMapping
+    public String mostrarAmenidades(HttpSession session, Model model) {
+        if (!cargarDatos(session, model)) {
             return "redirect:/login";
         }
 
-        model.addAttribute("isResidente", tipo.equals("RESIDENTE"));
-        model.addAttribute("isSeguridad", tipo.equals("SEGURIDAD"));
+        model.addAttribute("amenidadForm", new Amenidad());
+        model.addAttribute("tab", "consultar");
 
         return "Amenidades";
     }
 
     @GetMapping("/nueva")
-    public String nuevaAmenidad(Model model) {
-        model.addAttribute("amenidades", amenidadService.getAllAmenidad());
+    public String nuevaAmenidad(HttpSession session, Model model) {
+        if (!cargarDatos(session, model)){
+            return "redirect:/login";
+        }
         model.addAttribute("amenidadForm", new Amenidad());
         model.addAttribute("tab", "registrar");
+
         return "Amenidades";
     }
 
-    // GUARDAR (CORREGIDO)
     @PostMapping("/guardar")
-    public String guardar(@Valid @ModelAttribute("amenidadForm") Amenidad amenidad, BindingResult result, Model model) {
+    public String guardar(@Valid @ModelAttribute("amenidadForm") Amenidad amenidad, BindingResult result, HttpSession session, Model model) {
+        if (!cargarDatos(session, model)){
+            return "redirect:/login";
+        }
+
         if (result.hasErrors()) {
-            model.addAttribute("amenidades", amenidadService.getAllAmenidad());
             model.addAttribute("tab", "registrar");
             return "Amenidades";
         }
@@ -57,7 +68,6 @@ public class AmenidadController {
         try {
             amenidadService.saveAmenidad(amenidad);
         } catch (RuntimeException e) {
-            model.addAttribute("amenidades", amenidadService.getAllAmenidad());
             model.addAttribute("errorGeneral", e.getMessage());
             model.addAttribute("tab", "registrar");
             return "Amenidades";
@@ -66,28 +76,31 @@ public class AmenidadController {
         return "redirect:/amenidades";
     }
 
-    // EDITAR
     @GetMapping("/editar/{id}")
-    public String editar(@PathVariable Integer id, Model model) {
+    public String editar(@PathVariable Integer id, HttpSession session, Model model) {
+       if (!cargarDatos(session, model)){
+           return "redirect:/login";
+       }
+
         Amenidad amenidad = amenidadService.getAmenidadById(id);
-        model.addAttribute("amenidades", amenidadService.getAllAmenidad());
+
         model.addAttribute("amenidadForm", amenidad);
         model.addAttribute("tab", "editar");
         return "Amenidades";
     }
 
-    // ACTUALIZAR
     @PostMapping("/actualizar/{id}")
-    public String actualizar(@PathVariable Integer id,
-                             @Valid @ModelAttribute("amenidadForm") Amenidad amenidad,
-                             BindingResult result,
-                             Model model) {
+    public String actualizar(@PathVariable Integer id, @Valid @ModelAttribute("amenidadForm") Amenidad amenidad,
+                             BindingResult result, HttpSession session, Model model) {
 
-        // 1. Verificar si hay errores de validación
+        if (!cargarDatos(session, model)){
+            return "redirect:/login";
+        }
+
+        amenidad.setIdAmenidad(id);
+
         if (result.hasErrors()) {
-            amenidad.setIdAmenidad(id); // Asegúrate que el objeto mantenga su ID
-            model.addAttribute("amenidades", amenidadService.getAllAmenidad());
-            model.addAttribute("tab", "editar");
+           model.addAttribute("tab", "editar");
             return "Amenidades";
         }
 
@@ -95,34 +108,107 @@ public class AmenidadController {
             amenidadService.updateAmenidad(id, amenidad);
         } catch (RuntimeException e) {
             amenidad.setIdAmenidad(id);
-            model.addAttribute("amenidades", amenidadService.getAllAmenidad());
             model.addAttribute("errorGeneral", e.getMessage());
             model.addAttribute("tab", "editar");
             return "Amenidades";
         }
 
-        return "redirect:/Amenidades";
+        return "redirect:/amenidades";
     }
 
-    // ELIMINAR
     @PostMapping("/eliminar/{id}")
-    public String eliminar(@PathVariable Integer id) {
+    public String eliminar(@PathVariable Integer id, HttpSession session) {
+        String tipo = (String) session.getAttribute("tipoUsuario");
+
+        if (tipo == null){
+            return "redirect:/login";
+        }
+
         amenidadService.deleteAmenidad(id);
-        return "redirect:/Amenidades";
+
+        return "redirect:/amenidades";
     }
 
-    // BUSCAR
     @GetMapping("/buscar")
-    public String buscar(@RequestParam Integer id, Model model) {
-       try {
+    public String buscar(@RequestParam Integer id, HttpSession session, Model model) {
+       if (!cargarDatos(session, model)){
+           return "redirect:/login";
+       }
+
+        try {
            Amenidad amenidad = amenidadService.getAmenidadById(id);
            model.addAttribute("amenidades" , List.of(amenidad));
        } catch (Exception e) {
            model.addAttribute("amenidades", List.of());
            model.addAttribute("errorGeneral", "No se encontró la reservación con ID: " + id);
        }
+
        model.addAttribute("amenidadForm", new Amenidad());
        model.addAttribute("tab", "consultar");
        return "Amenidades";
+    }
+
+    @GetMapping("/residente")
+    public String vistaAmenidadesResidente(HttpSession session, Model model) {
+        if (!cargarDatos(session, model)){
+            return "redirect:/login";
+        }
+
+        model.addAttribute("amenidadForm", new Amenidad());
+        model.addAttribute("amenidadSeleccionada", null);
+        model.addAttribute("mostrarFormulario", false);
+
+        return "AmenidadesResidente";
+    }
+
+    @GetMapping("/residente/reservar")
+    public String mostrarFormularioReserva(@RequestParam String nombre, HttpSession session, Model model) {
+
+        if (!cargarDatos(session, model)){
+            return "redirect:/login";
+        }
+
+        Amenidad amenidad = new Amenidad();
+
+        Integer idResidente = (Integer) session.getAttribute("idResidente");
+
+        amenidad.setIdResidente(idResidente);
+        amenidad.setNombreAmenidad(nombre);
+        amenidad.setEstado("pendiente");
+
+        model.addAttribute("amenidadForm", amenidad);
+        model.addAttribute("amenidadSeleccionada", nombre);
+        model.addAttribute("mostrarFormulario", true);
+
+        return "AmenidadesResidente";
+    }
+
+    @PostMapping("/residente/guardar/amenidad")
+    public String guardarReservaResidente(@Valid @ModelAttribute("amenidadForm") Amenidad amenidad, BindingResult result, HttpSession session, Model model) {
+        if (!cargarDatos(session, model)){
+            return "redirect:/login";
+        }
+
+        Integer idResidente = (Integer) session.getAttribute("idResidente");
+
+        amenidad.setIdResidente(idResidente);
+        amenidad.setEstado("pendiente");
+
+        if (result.hasErrors()) {
+            model.addAttribute("amenidadSeleccionada", amenidad.getNombreAmenidad());
+            model.addAttribute("mostrarFormulario", true);
+            return "AmenidadesResidente";
+        }
+
+        try {
+            amenidadService.saveAmenidad(amenidad);
+        } catch (RuntimeException e) {
+            model.addAttribute("errorGeneral", e.getMessage());
+            model.addAttribute("amenidadSeleccionada", amenidad.getNombreAmenidad());
+            model.addAttribute("mostrarFormulario", true);
+            return "AmenidadesResidente";
+        }
+
+        return "redirect:/amenidades/residente";
     }
 }
